+++ sdk/search/azure-mgmt-search/generated_samples/search_create_or_update_service_with_identity.py
@@ -50,6 +50,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-03-01-preview/SearchCreateOrUpdateServiceWithIdentity.json
+# x-ms-original-file: 2026-09-01-preview/SearchCreateOrUpdateServiceWithIdentity.json
 if __name__ == "__main__":
     main()