+++ sdk/search/azure-mgmt-search/tests/test_search_management_usages_operations_test.py
@@ -8,9 +8,7 @@
 import pytest
 from azure.mgmt.search import SearchManagementClient
 
-from devtools_testutils import AzureMgmtRecordedTestCase, RandomNameResourceGroupPreparer, recorded_by_proxy
-
-AZURE_LOCATION = "eastus"
+from devtools_testutils import AzureMgmtRecordedTestCase, recorded_by_proxy
 
 
 @pytest.mark.live_test_only
@@ -18,9 +16,8 @@ class TestSearchManagementUsagesOperations(AzureMgmtRecordedTestCase):
     def setup_method(self, method):
         self.client = self.create_mgmt_client(SearchManagementClient)
 
-    @RandomNameResourceGroupPreparer(location=AZURE_LOCATION)
     @recorded_by_proxy
-    def test_usages_list_by_subscription(self, resource_group):
+    def test_usages_list_by_subscription(self):
         response = self.client.usages.list_by_subscription(
             location="eastus",
         )